Provide an adequate solution
Every example is tested with a JUnit test in place to verify it works. The String reader method is also verified to confirm that a file's contents can be read. Two different driver options are available: a graphical one and a none graphical one. By default the graphical one will always load, however if anyone wants to try out the console-based one they are more than welcome to do so.